backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 环境变量
python-dotenv==1.0.0

# HTTP客户端
httpx==0.26.0
openai==1.7.0

# 工具库
python-multipart==0.0.6

# JSON处理
orjson==3.9.12

# YAML配置文件
PyYAML==6.0.1
